// maretinda-store/src/components/molecules/LoginForm/schema.ts
import { z } from 'zod';

export const loginFormSchema = z.object({
	email: z.string().nonempty('Please enter email').email('Invalid email'),
	password: z.string().nonempty('Please enter password'),
	remember: z.boolean().optional(),
});

export type LoginFormData = z.infer<typeof loginFormSchema>;